pyproject.toml
[project]
name = "sinapsis-retina-face-trt"
version = "0.1.6"
description = "trt version of retina face and facenet"
authors = [
    {name = "SinapsisAI", email = "dev@sinapsis.tech"},
]
readme = "README.md"
license-files= ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
    "packaging>=24.2",
    "polars>=1.26.0",
    "retinaface-pytorch>=0.0.7",
    "sinapsis>=0.2.3",
    "sinapsis-data-readers>=0.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.8.3",
    "pre-commit>=4.0.1",
]
[project.optional-dependencies]
model-conversion = [
    "deepface==0.0.92",
    "sinapsis-framework-converter[all]>=0.1.0",
]
webapp-gradio = [
    "sinapsis-retina-face-trt[sinapsis-data-visualization]",
    "sinapsis[webapp]>=0.2.3",
]
sinapsis-data-visualization = [
    "sinapsis-data-visualization[visualization-matplotlib]>=0.1.0",
]
sinapsis-data-writers = [
    "sinapsis-data-writers>=0.1.1",
]
all = [
    "sinapsis-retina-face-trt[dev,model-conversion,sinapsis-data-visualization,sinapsis-data-writers,webapp-gradio]",
]


[tool.ruff]
lint.select = [
    "ARG",   # Unused arguments
    "ANN",
    "BLE",   # Blind except
    "C4",    # Comprehensions
    "E",     # Errors
    "F",     # Pyflakes
    "FIX",   # "FIXME-comments"
    "FLY",   # F-strings
    "I",     # Isort
    "PERF",  # Unnecessary performance costs
    "PIE",   # Unnecessary code
    "RUF",   # Ruff
    "RSE",   # Raise
    "SIM",   # Code simplification
    "SLOT",  # __slots__
    "T10",   # Debugger
    "T20",   # Print
    "TD",    # "TODO-comments"
    "TID",   # Tidy imports
]
lint.ignore = ['ANN401']
line-length = 120
show-fixes = true

[[tool.uv.index]]
url = "https://pypi.sinapsis.tech/"

[tool.uv.sources]
sinapsis-retina-face-trt = { workspace = true }

[project.urls]
Homepage = "https://sinapsis.tech"
Documentation = "https://docs.sinapsis.tech/docs"
Tutorials = "https://docs.sinapsis.tech/tutorials"
Repository = "https://github.com/Sinapsis-AI/sinapsis-retina-face-trt.git"
